refactor: Remove GPU memory pool and add TensorRT L2 cache configuration
- Remove CudaMemoryPool class and related code (unnecessary with CUDA Graph)
- Add persistent_cache_limit_mb configuration for TensorRT L2 cache optimization
- Update UI to allow L2 cache size adjustment (1-64MB range)
- Fix size_t to int conversion warnings in simple_capture.cpp
- Simplify GPU resource management by relying on CUDA Graph's built-in memory optimization

+++ needaimbot/capture/simple_capture.cpp
@@ -140,7 +140,7 @@ SimpleCudaMat SimpleScreenCapture::GetNextFrameGpu()
     cuda_bgra2bgr(
         m_deviceFrame.data(), gpuFrame.data(),
         m_width, m_height,
-        m_deviceFrame.step(), gpuFrame.step()
+        static_cast<int>(m_deviceFrame.step()), static_cast<int>(gpuFrame.step())
     );
     
     return gpuFrame;
